[Updated] Task Management with some extra validation

+++ Projects/Task Management/Services/TaskRepoServices.cs
@@ -8,18 +8,19 @@ namespace Task_Management.Services
     public class TaskRepoServices : ITaskRepository
     {
         private readonly ApplicationDbContext _context;
+        private readonly string UserId;
 
         public TaskRepoServices(ApplicationDbContext context)
         {
             _context = context;
+            HttpContext httpContext = new HttpContextAccessor().HttpContext;
+            UserId = httpContext.User.FindFirstValue(ClaimTypes.NameIdentifier);
         }
 
         public void CreateTask(Models.Task task)
         {
-            HttpContext httpContext = new HttpContextAccessor().HttpContext;
-            var userId = httpContext.User.FindFirstValue(ClaimTypes.NameIdentifier);
-            task.UserId = userId;
-            task.User = _context.Users.Find(userId);
+            task.UserId = UserId;
+            task.User = _context.Users.Find(UserId);
             _context.Tasks.Add(task);
             _context.SaveChanges();
         }
@@ -27,7 +28,7 @@ namespace Task_Management.Services
         public void DeleteTask(int id)
         {
             Models.Task task = _context.Tasks.Find(id);
-            if (task != null)
+            if (task != null && UserId == task.UserId)
             {
                 _context.Tasks.Remove(task);
                 _context.SaveChanges();
@@ -36,14 +37,18 @@ namespace Task_Management.Services
 
         public List<Models.Task> GetAllTasks()
         {
-            HttpContext httpContext = new HttpContextAccessor().HttpContext;
-            var userId = httpContext.User.FindFirstValue(ClaimTypes.NameIdentifier);
-            return _context.Tasks.Where(x => x.UserId == userId).ToList();
+            
+            return _context.Tasks.Where(x => x.UserId == UserId).ToList();
         }
 
         public Models.Task GetTaskById(int id)
         {
-            return _context.Tasks.Find(id);
+            var task = _context.Tasks.Where(x => x.UserId == UserId && x.Id == id);
+            if(task != null)
+                return task.FirstOrDefault();
+            else
+                return null;
+
 
         }
 
@@ -51,7 +56,7 @@ namespace Task_Management.Services
         public void UpdateTask(Models.Task task, int id)
         {
             Models.Task updatedTask = _context.Tasks.FirstOrDefault(x => x.Id == id);
-            if (updatedTask != null)
+            if (updatedTask != null && updatedTask.UserId == UserId)
             {
                 updatedTask.TaskTime = task.TaskTime;
                 updatedTask.TaskName = task.TaskName;
@@ -59,6 +64,7 @@ namespace Task_Management.Services
                 updatedTask.TaskDescription = task.TaskDescription;
                 _context.SaveChanges();
             }
+            
         }
